refactor(tranche-1): remove dead code and fix black-color cancel ambiguity
- Inline upstream_ignoring_hidden into anchor.py and delete util.py
  (drop orphaned select_upstream_ignoring_hidden; remove lazy import)
- Delete _offer_make_dot_anchor in anchor.py (deprecated, never called
  from production); drop tautological patch in test_dot_type_distinction.py
- Fix black-as-cancel ambiguity in colors.py custom-color picker at all
  three sites (lines 411, 1106, 1123): use sentinel default instead of
  treating result == 0 as cancel; update related test to match.

Behaviour preserved except the black-color bug fix (user-confirmed).
Part of a 5-tranche refactor; see docs/plans/refactor_simplify_codebase.md.
